% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimVarParam.R
\name{ClimVarParam}
\alias{ClimVarParam}
\title{Subfunction for obtaining climatological values of sill, range, and nugget of spherical variogram model for RAINLINK.}
\usage{
ClimVarParam(DateStr)
}
\arguments{
\item{DateStr}{the end date of the chosen daily period.}
}
\value{
Sill, Range and Nugget
}
\description{
Subfunction for obtaining climatological values of sill, range, and nugget of spherical variogram model. This is based on a climatological variogram based on 30-year automatic rain gauge data sets from The Netherlands. Spherical variograms have been modelled as function of the day number and duration in Van de Beek et al. (2012). They use durations of 1 - 24 h. In this function the relationships can be extrapolated to, e.g. 15-min, data. For all input data the middle of the link is used as location of the path-averaged measurement. Input data are link-based, e.g. 15-minute, path-averaged rainfall accumulations. These rainfall accumulations are assumed to be representative for the point at the center of the link. Van de Beek, C. Z., Leijnse, H., Torfs, P. J. J. F., and Uijlenhoet, R.: Seasonal semi-variance of Dutch rainfall at hourly to daily scales, Adv. Water Resour., 45, 76-85, doi:10.1016/j.advwatres.2012.03.023, 2012.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
}
}
\examples{
-

}

