% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectMinMaxRSL.R
\name{CorrectMinMaxRSL}
\alias{CorrectMinMaxRSL}
\title{Function for correcting minimum and maximum received signal powers.}
\usage{
CorrectMinMaxRSL(Data, Dry, Pref)
}
\arguments{
\item{Data}{Data frame with microwave link data}

\item{Dry}{Data frame: Should interval be considered dry for reference level 
determination? (0 = wet; 1 = dry)}

\item{Pref}{Reference level (dB)}
}
\value{
Data frame with corrected minimum and maximum received powers (dB)
}
\description{
Function for correcting minimum and maximum received signal powers.
For a rainy time interval the corrected minimum received signal power becomes equal to 
the minimum received signal power if this is below the reference signal level. 
Otherwise the corrected minimum received signal power becomes equal to the 
reference signal level. The corrected maximum received signal power becomes
equal to the maximum received signal power if both the maximum received signal 
power and the corrected minimum received signal power are below the reference
signal level.

Works for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.
}
\examples{
CorrectMinMaxRSL(Data=DataOutlierFiltered,Dry=WetDry$Dry,Pref=Pref)
}
\author{
Aart Overeem & Hidde Leijnse
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R. (2016): Retrieval algorithm for rainfall mapping from
microwave links in a cellular communication network, Atmospheric Measurement Techniques, under review.
}

