% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAvailability.R
\name{DataAvailability}
\alias{DataAvailability}
\title{Function for plotting data availability of commercial microwave link network.}
\usage{
DataAvailability(
  Data,
  cex.axis,
  cex.lab,
  FigNameBarplotAvailabilityLinks,
  FigNameBarplotAvailabilityLinkPaths,
  FigNameTimeseriesAvailability,
  ps,
  Rmean,
  TimeZone = "UTC",
  LocationLegendTimeseriesAvailability,
  MaxPercLinkPaths,
  MaxPercSubLinks,
  PlotTitle,
  verbose = TRUE
)
}
\arguments{
\item{Data}{Data frame with microwave link data (use data(Linkdata) to load example data).}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of ''cex'' in file FigNameTimeseriesAvailability.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of ''cex'' in file FigNameTimeseriesAvailability.}

\item{FigNameBarplotAvailabilityLinks}{Name of file with bar plot with percentage of links for bins of link orientation. The extension must be ''.pdf''.}

\item{FigNameBarplotAvailabilityLinkPaths}{Name of file with bar plot with percentage of links for bins of link path length. The extension must be ''.pdf''.}

\item{FigNameTimeseriesAvailability}{Name of file with bar plot with percentage of links for bins of microwave frequency. The extension must be ''.pdf''.}

\item{LocationLegendTimeseriesAvailability}{Location of legend of figure with time series of number of sub-links and link paths. Choose from "bottomright", "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center"}

\item{MaxPercLinkPaths}{Maximum percentage on scale for bar plot of link path availability.}

\item{MaxPercSubLinks}{Maximum percentage on scale for bar plot of sub-link availability.} 

\item{PlotTitle}{Title of plots (e.g., which CML vendor or period).}

\item{ps}{integer; the point size of text (but not symbols) in file FigNameTimeseriesAvailability.}

\item{Rmean}{Vector of link-derived rainfall intensities (mm h\eqn{^{-1}}) with length equal to Data.}

\item{TimeZone}{Time zone of data (e.g. "UTC").}
}
\value{
Figures with data availability of commercial microwave link network.
}
\description{
Function for plotting data availability of commercial microwave link data. This function provides the following figures:  
\enumerate{
  \item Bar plot with percentage of sub-links for bins of data availability (%).
  \item Bar plot with percentage of link paths for bins of data availability (%).
  \item Plot with average number of sub-links and link paths for each time interval of the sampling strategy.
}
In addition, the Average number of sub-links and link paths is printed to the screen.

Does not depend on sampling strategy.

The input microwave link data do not have to be sorted chronologically. Full-duplex links will give two data entries, these will both be used.

Data availability is computed over the number of time intervals in the dataset as obtained from DateTime.
In case a time interval is missing, i.e. not present in DateTime, it is not included in the computations.
Hence, the computed availability does only refer to the time intervals for which a DateTime entry is present in the dataset.
In the computation of availabilities only links or paths are counted as available in case the link-derived rainfall intensities
are equal to or larger than 0 (mm h\eqn{^{-1}}). I.e., NA values in Rmean are taken into account in the computation of data availability.
So data availability is always computed over all time intervals in DateTime. Note that NA values do not exist in the DateTime object if function
''PreprocessingMinMaxRSL'' has been run and note that Data must be preprocessed, because Rmean is used.

New as of RAINLINK version 1.31: When Rmean is not available, just use Rmean=NULL as argument. Then, the availability is also computed, but no selections with Rmean are performed.
So no preprocessing is needed in that case.
}
\examples{
data(Linkdata)
DataAvailability(Data=Linkdata,cex.axis=0.9,cex.lab=1.15,
FigNameBarplotAvailabilityLinks="Barplot_Availability_Links.pdf",
FigNameBarplotAvailabilityLinkPaths="Barplot_Availability_LinkPaths.pdf",
FigNameTimeseriesAvailability="TimeseriesAvailability.pdf",ps=18,Rmean=Rmean,TimeZone="UTC",
MaxPercLinkPaths=100,MaxPercSubLinks=100,PlotTitle="Timeseries of data availability",
LocationLegendTimeseriesAvailability="topright")
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem
}
