% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDW.R
\name{IDW}
\alias{IDW}
\title{Subfunction for inverse distance weighted interpolation on point data.}
\usage{
IDW(rain.grid)
}
\value{
InterpField
}
\description{
Subfunction for inverse distance weighted interpolation on point data. 
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item idp: The inverse distance weighting power.
}
}
\examples{
-

}

