% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{Interpolation}
\alias{Interpolation}
\title{Interpolation of link-based path-averaged rainfall estimates.}
\usage{
Interpolation()
}
\value{
InterpField
}
\description{
Interpolation of link-based path-averaged rainfall estimates. The type of interpolation can be specified in the configuration file. The following types are available: 1) Inverse distance weighted interpolation on data (subfunction IDW); 2) Ordinary kriging with spherical variogram model. Its parameter values nugget, sill, and range, can be defined by the user; 3) Ordinary kriging with spherical variogram model with climatological parameter values based on a 30-year rain gauge data set. These are computed for the day of year as obtained from the file name, thus taking into account seasonality in spatial rainfall correlation. The subfunction ClimVarParam computes these parameter values. Ordinary kriging is performed by subfunction OrdinaryKriging. Note that this interpolation algorithm is developed for interpolation of link-based rainfall estimates, which are path averages. The subfunction IntpPathToPoint computes the path-averaged rainfall intensity for unique link paths. And it assigns path-averaged intensity to the point at the middle of the link path.
\itemize{
  \item FolderRainEstimates: Folder name of input data.
  \item FolderRainMaps: Folder name of output data.
  \item CoorSystemInputData: Coordinate system of input data (e.g. "+init=epsg:4326"	for WGS84 in degrees).
  \item FileGrid: File with interpolation grid.
  \item IntpMethod: Ordinary kriging ("OK") or inverse distance weighted interpolation ("IDW").
  \item ClimVar: Use "ClimvdBeek" for climatological spherical variogram model. Use "Manual" for spherical variogram model with nugget, sill, and range values manually specified.
  \item projstring: Proj4string of an Azimuthal Equidistant Cartesian output coordinate system.
}
}
\examples{
-

}

