% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntpPathToPoint.R
\name{IntpPathToPoint}
\alias{IntpPathToPoint}
\title{Subfunction for computing path-averaged rainfall intensity for unique link paths. Assign the path-averaged intensity to point at middle of link path.}
\usage{
IntpPathToPoint(Data)
}
\arguments{
\item{Data}{the microwave link data}
}
\value{
Rainlink, NrPaths, NrLinks
}
\description{
Subfunction for computing path-averaged rainfall intensity for unique link paths. Assign the path-averaged intensity to point at middle of link path.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item ConversionDepthToIntensity: Conversion factor from rainfall depth to intensity.
}
}
\examples{
-

}

