% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxRSLToMean.R
\name{MinMaxRSLToMeanR}
\alias{MinMaxRSLToMeanR}
\title{Subfunction for path-averaged rainfall estimation from minimum and maximum attenuations 
from microwave links.}
\usage{
MinMaxRSLToMeanR(a, Aa, alpha, b, PathLength, PmaxCor, PminCor, Pref)
}
\arguments{
\item{a}{Coefficients in relationship between rainfall intensity and specific 
attenuation (mm h\eqn{^{-1}} dB\eqn{^{-b}} km\eqn{^{b}})}

\item{Aa}{Wet antenna attenuation correction \eqn{A_{\mbox{a}}} (dB)}

\item{alpha}{Coefficient (\eqn{\alpha}) determining contribution of minimum and 
maximum path-averaged rainfall intensity to mean path-averaged rainfall intensity (-)}

\item{b}{Exponents in relationship between rainfall intensity and specific attenuation (-)}

\item{PathLength}{Lengths of link paths (km)}
}
\value{
Data frame with mean path-averaged rainfall intensities (mm h\eqn{^{-1}})
}
\description{
Subfunction for path-averaged rainfall estimation using microwave links. 
Compute minimum and maximum attenuation over the link path. Convert these to minimum and 
maximum path-averaged rainfall intensities. Convert minimum and maximum path-averaged 
rainfall intensities to mean path-averaged rainfall intensities. 

Works for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.
}
\examples{
MinMaxRSLToMeanR(a=a,Aa=Aa,alpha=alpha,b=b,PathLength=Data$PathLength,
PmaxCor=PmaxCor,PminCor=PminCor,Pref=Pref)
}
\author{
Aart Overeem & Hidde Leijnse
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R. (2016): Retrieval algorithm for rainfall mapping from
microwave links in a cellular communication network, Atmospheric Measurement Techniques, under review.
}

