% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxToMean.R
\name{MinMaxToMean}
\alias{MinMaxToMean}
\title{Subfunction for path-averaged rainfall estimation using microwave links from minimum and maximum attenuations over link path.}
\usage{
MinMaxToMean()
}
\value{
Rmean
}
\description{
Subfunction for path-averaged rainfall estimation using microwave links. Convert minimum and maximum path-averaged rainfall intensity to mean path-averaged rainfall intensity. Works for sampling strategy where minimum and maximum received powers are provided.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item Aa: Wet antenna attenuation correction (dB).
  \item alpha: Parameter for conversion from minimum and maximum to mean path-averaged rainfall intensity.
}
}
\examples{
-

}

