% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoWetDryMinMax.R
\name{NoWetDryMinMax}
\alias{NoWetDryMinMax}
\title{Function for determination of reference signal level, and computing corrected received powers.}
\usage{
NoWetDryMinMax()
}
\description{
Function for determination of reference signal level (subfunction RefLevelMinMax), and computing corrected received powers, and corrected received powers (this function and subfunction RefLevelMinMax). No wet-dry classification of time intervals is applied. Further, the minimum and maximum attenuation over a link path are computed.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
  \item FolderPreprocessed: Folder name of input data.
  \item FolderCorrected: Folder name of output data.
  \item FileRainRetr: Filename with values of coefficients in relationship between specific attenuation and rainfall intensity.
}
}
\examples{
-

}

