% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinaryKriging.R
\name{OrdinaryKriging}
\alias{OrdinaryKriging}
\title{Subfunction for ordinary kriging interpolation of point values using spherical variogram model with predefined parameters sill, range, and nugget.}
\usage{
OrdinaryKriging(rain.grid)
}
\arguments{
\item{rain.grid}{rain grid and rainfall intensities}
}
\value{
InterpField
}
\description{
Subfunction for ordinary kriging interpolation of point values using spherical variogram model with predefined parameters sill, range, and nugget.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item nmax: For local kriging: the number of nearest observations that should be used for a kriging prediction or simulation, where nearest is defined in terms of the space of the spatial locations. By default, all observations are used.
}
}
\examples{
-

}

