% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFilterMinMax.R
\name{OutlierFilterMinMax}
\alias{OutlierFilterMinMax}
\title{Subfunction to apply filter to remove outliers in path-averaged microwave link attenuations.}
\usage{
OutlierFilterMinMax(Data)
}
\arguments{
\item{Data}{the microwave link data}
}
\value{
ID, a, b, Amax, Amin, PathLength, IntervalNumber, XStart, YStart, XEnd, YEnd, Frequency
}
\description{
Subfunction to apply filter to remove outliers in link-based rainfall estimates. Malfunctioning link antennas can cause outliers in rainfall retrievals (especially for daily accumulations). These outliers can be removed by using a filter that is based on the assumption that rainfall is correlated in space. The filter discards a time interval of a link for which the cumulative difference between its specific attenuation and that of the surrounding links over the previous 24 h (including the present time interval) becomes lower than -32.5 dB h per km. Works for sampling strategy where minimum and maximum received powers are provided.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item OUTLIERFILTER: Threshold outlier filter (-32.5 dB h per km).
}
}
\examples{
-

}

