% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFilterMinMaxRSL.R
\name{OutlierFilterMinMaxRSL}
\alias{OutlierFilterMinMaxRSL}
\title{Function to apply filter to remove outliers in path-averaged microwave link attenuations.}
\usage{
OutlierFilterMinMaxRSL(Data, F, FilterThreshold = -32.5)
}
\arguments{
\item{Data}{Data frame with microwave link data.}

\item{F}{Values for filter to remove outliers (dB km\eqn{^{-1}} h).}

\item{FilterThreshold}{Outlier filter threshold (dB h km\eqn{^{-1}}).}
}
\value{
Data frame with microwave link data.
}
\description{
Function to apply filter to remove outliers in link-based rainfall estimates. 
Malfunctioning link antennas can cause outliers in rainfall retrievals (especially for 
daily accumulations). These outliers can be removed by using a filter that is based on the 
assumption that rainfall is correlated in space. The filter discards a time interval of a 
link for which the cumulative difference between its specific attenuation and that of the 
surrounding links over the previous 24 h (including the present time interval), F, becomes 
lower than a threshold value in dB h km\eqn{^{-1}}. 

Works for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.

The outlier filter has been extensively tested on minimum received signal powers, i.e.
for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.
This function can also be applied in case of other sampling strategies, because
it does not explicitly require minimum and maximum received signal powers.
It just applies the selection on all rows in a data frame.
Whether the outlier filter will give good results when applied to link data
obtained from other sampling strategies would need to be tested.
Hence, ''MinMaxRSL'' is kept in this function name to stress that it
has been tested for a sampling strategy where minimum and maximum received 
powers are provided.

Also works for a sampling strategy where instantaneous transmitted and received signal levels are obtained.
In case of instantaneous signal levels, it does not matter whether transmitted power levels vary or are constant.
The only requirement is that the input data for RAINLINK needs some preprocessing. See ''ManualRAINLINK.pdf''
for instructions. 

Also works for a sampling strategy where average transmitted and received signal levels are obtained.
In case of average signal levels, it does not matter whether transmitted power levels vary or are constant.
The only requirement is that the input data for RAINLINK needs some preprocessing. See ''ManualRAINLINK.pdf''
for instructions. 

Can only be applied when function WetDryNearbyLinkApMinMaxRSL has been executed.
}
\examples{
OutlierFilterMinMaxRSL(Data=DataPreprocessed,F=WetDry$F,FilterThreshold=-32.5)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem & Hidde Leijnse
}
