% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathLength.R
\name{PathLength}
\alias{PathLength}
\title{Function to compute path length of microwave links from coordinates.}
\usage{
PathLength(InputCoorSystem, XStart, XEnd, YStart, YEnd)
}
\arguments{
\item{InputCoorSystem}{Define EPSG code for input coordinate system (e.g., 4326L for WGS84 in degrees).}

\item{XStart}{List with longitude of start of microwave links.}

\item{XEnd}{List with longitude of end of microwave links.}

\item{YStart}{List with latitude of start of microwave links.}

\item{YEnd}{List with latitude of end of microwave links.}
}
\value{
Path length of microwave links (km).
}
\description{
Function for computing path length of microwave links from coordinates.
}
\examples{
PathLength(InputCoorSystem=4326L,XStart=XStart,XEnd=XEnd,YStart=YStart,YEnd=YEnd)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem
}
