% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLinkLocations.R
\name{PlotLinkLocations}
\alias{PlotLinkLocations}
\title{Function which visualises microwave link paths on a map.}
\usage{
PlotLinkLocations(AlphaLinkLocations, BBoxOSMauto, OSMBottom, ColourLinks,
  ColourType, dataf, DateTime, ExtraTextLinkLocations, FigFileLinkLocations,
  FigHeight, FigWidth, FolderFigures, FontFamily, GoogleLocDegSpecified,
  GoogleLocLat, GoogleLocLon, GoogleLocName, GoogleLocNameSpecified,
  GoogleMapType, GoogleZoomlevel, LabelAxisLat, LabelAxisLonGoogle,
  LabelAxisLonOSM, OSMLeft, MapBackground, OSMRight, OSMScale, SizeLinks,
  SizePlotTitle, TitleLinkLocations, OSMTop)
}
\arguments{
\item{AlphaLinkLocations}{Transparency of link paths}

\item{BBoxOSMauto}{Compute bounding box from input data or used bounding box defined above?
(for OpenStreetMap only). Use "yes" if bounding box is to be computed from interpolation grid.}

\item{OSMBottom}{Latitude in degrees (WGS84) for bottom side of the area for which rainfall depths 
are to be plotted (for OpenStreetMap only).}

\item{ColourLinks}{Colour of plotted link paths}

\item{ColourType}{Colour or black-and-white background map? Use "color" for colour and "bw" 
for black-and-white background map.}

\item{dataf}{data frame which contains (at least) locations of microwave links in Azimuthal 
Equidistant Cartesian coordinate system}

\item{DateTime}{Date and time for which link locations are plotted. This is used in the
title caption of the figure and in the file name}

\item{ExtraTextLinkLocations}{Second part of title of plot}

\item{FigFileLinkLocations}{Part of figure output file name}

\item{FigHeight}{Figure height. 1280 times 1280 pixels seems maximum graphical resolution for 
downloaded Google Maps. Because also axes and legend are plotted, it is advised to use e.g. 
1450 times 1450 pixels. 
Then the Google Map will remain 
approximately 1280 times 1280 pixels. Using higher values is not a problem (e.g. 2000). 
In this way it is tried to get the highest possible resolution. For OpenStreetMap the maps may 
reach resolutions of 1500 - 2000 pixels. Hence, using FigWidth and FigHeight of 2000 pixels or 
higher is advised. The OpenStreetMap itself is stored in file "ggmapTemp.png". From this file 
the resolution of the background map can be obtained. This can be useful for determining an 
appropriate FigWidth and FigHeight above.}

\item{FigWidth}{Figure width. 1280 times 1280 pixels seems maximum graphical resolution for 
downloaded Google Maps. Because also axes and legend are plotted, it is advised to use e.g. 
1450 times 1450 pixels. Then the Google Map will remain approximately 1280 times 1280 pixels. 
Using higher values is not a problem (e.g. 2000). In this way it is tried to get the 
highest possible resolution. For OpenStreetMap the maps may reach resolutions of 1500 - 2000 
pixels. Hence, using FigWidth and FigHeight of 2000 pixels or higher is advised. The OpenStreetMap 
itself is stored in file "ggmapTemp.png". From this file the resolution of the background map can 
be obtained. This can be useful for determining an appropriate FigWidth and FigHeight above.}

\item{FolderFigures}{Folder name of figures}

\item{FontFamily}{Specify font family of text in figures. To select the default font use "". 
Using "Times" may give warnings when executing the visualisation. In that case the font is not 
installed on the computer. This can be solved by using the default font ("").}

\item{GoogleLocDegSpecified}{If GoogleLocDegSpecified is "yes" then the specified location in 
degrees is used is used as the centre of the Google Map. If both GoogleLocNameSpecified and 
GoogleLocDegSpecified are not equal to "yes", the bounding box of the map is determined from 
the provided grid and used as centre of the Google Map.}

\item{GoogleLocLat}{Latitude of middle of Google Map (degrees)}

\item{GoogleLocLon}{Longitude of middle of Google Map (degrees).}

\item{GoogleLocName}{Location of middle of Google Map, provided as text, e.g. name of city,
street name, country}

\item{GoogleLocNameSpecified}{If GoogleLocNameSpecified is "yes" then the specified location 
name GoogleLocName is used as the centre of the Google Map. If both GoogleLocNameSpecified and 
GoogleLocDegSpecified are not equal to "yes", the bounding box of the map is determined from the 
provided grid and used as centre of the Google Map.}

\item{GoogleMapType}{In case of Google Maps: which map type should be used? Available map types: 
"terrain", "satellite", "roadmap", and "hybrid".}

\item{GoogleZoomlevel}{Which zoom level to use for the Google Maps?}

\item{LabelAxisLat}{Label name of vertical axis}

\item{LabelAxisLonGoogle}{Label name of horizontal axis (for Google Maps only)}

\item{LabelAxisLonOSM}{Label name of horizontal axis (for OpenStreetMap only)}

\item{OSMLeft}{Longitude in degrees (WGS84) for left side of the area for which rainfall depths are 
to be plotted (for OpenStreetMap only).}

\item{MapBackground}{Google Maps or OpenStreetMap as background? Use "Google" for Google Maps and 
"OSM" for OpenStreetMap. Note that Google Maps will only plot on a square figure.}

\item{OSMRight}{Longitude in degrees (WGS84) for right side of the area for which rainfall depths 
are to be plotted (for OpenStreetMap only).}

\item{OSMScale}{Give value of scale (for OpenStreetMap only). A proper choice of the scale parameter 
in get_openstreetmap is difficult. It cannot be computed automatically. Hence, a scale parameter value 
should be provided below. The scale parameter should be as small as possible to get the highest 
graphical resolution. However, a too low value may result in a map not being downloaded. Hence, the user 
should manually supply get_openstreetmap with a scale. It may require some iterations to find the 
appropriate value for scale. The file "ggmapTemp.png" is written to disk when an OpenStreetMap is loaded. 
The highest possible resolution for a square area is about 2000 x 2000 pixels.}

\item{SizeLinks}{Size of plotted link paths}

\item{SizePlotTitle}{Size of plot title}

\item{TitleLinkLocations}{First part of title of plot}

\item{OSMTop}{Latitude in degrees (WGS84) for top side of the area for which rainfall depths are to be 
plotted (for OpenStreetMap only)}
}
\description{
Function which visualises microwave link paths on a map.
}
\examples{
PlotLinkLocations(AlphaLinkLocations=AlphaLinkLocations,
BBoxOSMauto=BBoxOSMauto,OSMBottom=OSMBottom,
ColourLinks=ColourLinks,ColourType=ColourType,
dataf=dataf,DateTime=DateTime,ExtraTextLinkLocations=ExtraTextLinkLocations,
FigFileLinkLocations=FigFileLinkLocations,FigHeight=FigHeight,
FigWidth=FigWidth,FolderFigures=FolderFigures,FontFamily=FontFamily,
GoogleLocDegSpecified=GoogleLocDegSpecified,GoogleLocLat=GoogleLocLat,
GoogleLocLon=GoogleLocLon,GoogleLocName=GoogleLocName,
GoogleLocNameSpecified=GoogleLocNameSpecified,
GoogleMapType=GoogleMapType,GoogleZoomlevel=GoogleZoomlevel,
LabelAxisLat=LabelAxisLat,LabelAxisLonGoogle=LabelAxisLonGoogle,
LabelAxisLonOSM=LabelAxisLonOSM,OSMLeft=OSMLeft,
MapBackground=MapBackground,OSMRight=OSMRight,OSMScale=OSMScale,
SizeLinks=SizeLinks,SizePlotTitle=SizePlotTitle,
TitleLinkLocations=TitleLinkLocations,OSMTop=OSMTop)
}
\author{
Aart Overeem & Hidde Leijnse
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R. (2016): Retrieval algorithm for rainfall mapping from
microwave links in a cellular communication network, Atmospheric Measurement Techniques, under review.
}

