% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygons.R
\name{Polygons}
\alias{Polygons}
\title{Subfunction which makes dataframe for polygons with rainfall estimates in specific rainfall class.}
\usage{
Polygons(cond, Selected)
}
\arguments{
\item{cond}{row numbers of dataframe which fall in specific rainfall class}

\item{Selected}{coordinates of polygons and their assigned rainfall values}
}
\value{
dataf, available
}
\description{
Subfunction which makes dataframe for polygons with rainfall estimates in specific rainfall class.
}
\examples{
-

}

