% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingMinMax.R
\name{PreprocessingMinMax}
\alias{PreprocessingMinMax}
\title{Function for preprocessing of microwave link data.}
\usage{
PreprocessingMinMax()
}
\description{
Function for preprocessing of microwave link data. Works for sampling strategy where minimum and maximum received powers are provided. This function performs the following tasks:  
\enumerate{
  \item Time interval numbers are computed.
  \item Link data are selected for microwave frequencies within chosen range.
  \item Data selection criteria are applied.
  \item Link coordinates are converted to Cartesian coordinate system. 
  \item Data from previous and present 24-h period are combined into one file for each day for which rainfall maps need to be obtained. 
}
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
  \item FolderStart: Folder name of input data.
  \item FolderPreprocessed: Folder name of output data.
  \item TimeZone: Time zone of data (e.g. "UTC").
  \item PERIOD: Select daily time interval, i.e., "0800" implies 0800 UTC previous day - 0800 UTC present day (use 2400 for 0000 UTC).
  \item MinFrequency: Minimum allowed microwave frequency of link in output (GHz).
  \item MaxFrequency: Maximum allowed microwave frequency of link in output (GHz).
  \item CoorSystemInputData: Coordinate system of input data (e.g. "+init=epsg:4326"	for WGS84 in degrees).
  \item projstring: Proj4string of an Azimuthal Equidistant Cartesian output coordinate system.
}
}
\examples{
-

}

