% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingMinMaxRSL.R
\name{PreprocessingMinMaxRSL}
\alias{PreprocessingMinMaxRSL}
\title{Function for preprocessing of microwave link data.}
\usage{
PreprocessingMinMaxRSL(Data, MaxFrequency = Inf, MinFrequency = 0,
  verbose = TRUE)
}
\arguments{
\item{Data}{Data frame with microwave link data (use data(Linkdata) to load example data)}

\item{MaxFrequency}{Maximum allowed microwave frequency of link in output (GHz; default infinite)}

\item{MinFrequency}{Minimum allowed microwave frequency of link in output (GHz; default 0)}
}
\value{
Data frame with microwave link data
}
\description{
Function for preprocessing of microwave link data. This function performs the following tasks:  
\enumerate{
  \item Link data are selected for microwave frequencies within chosen range.
  \item Data selection criteria are applied.
}

Works for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.
}
\examples{
data(Linkdata)
PreprocessingMinMaxRSL(Data=Linkdata,MaxFrequency=40.5,MinFrequency=12.5)
}
\author{
Aart Overeem & Hidde Leijnse
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R. (2016): Retrieval algorithm for rainfall mapping from
microwave links in a cellular communication network, Atmospheric Measurement Techniques, under review.
}

