% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainMapsLinksDaily.R
\name{RainMapsLinksDaily}
\alias{RainMapsLinksDaily}
\title{Function which visualises daily link-based rainfall depths.}
\usage{
RainMapsLinksDaily(DateMap)
}
\arguments{
\item{DateMap}{end date of daily period for which rainfall map should be plotted}
}
\description{
Function which visualises daily link-based rainfall depths. Requires interpolation grid and file with polygons of pixels. Daily rainfall depths are computed irrespective of the number of available files. If, for instance, only one out of 96 files is available, the daily rainfall depth is still computed and visualised. The data availability is plotted as a percentage in the title caption of the graph.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
}
}
\examples{
RainMapsLinksDaily(20110911) # To plot daily link rainfall map ending at 11 September 2011.

}

