% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainMapsLinksTimeStep.R
\name{RainMapsLinksTimeStep}
\alias{RainMapsLinksTimeStep}
\title{Function which visualises link-based rainfall depths for chosen TIMESTEP in configuration file.}
\usage{
RainMapsLinksTimeStep()
}
\description{
Function which visualises link-based rainfall depths for chosen TIMESTEP in configuration file. Requires interpolation grid and file with polygons of pixels, the polygons having four angular points. Note that the zoom level for Google Maps can be chosen. This zoom level will not always match with the area to be plotted. Downloading a GoogleMaps background map is very fast, but downloading an OpenStreetMap may require tenths of seconds. Note that such a map is downloaded only once, which saves time in case of multiple rainfall maps. 
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
}
}
\examples{
RainMapsLinksTimeStep()

}

