% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainMapsRadarsDaily.R
\name{RainMapsRadarsDaily}
\alias{RainMapsRadarsDaily}
\title{Function which visualises daily radar rainfall depths.}
\usage{
RainMapsRadarsDaily(DateMap)
}
\arguments{
\item{DateMap}{end date of daily period for which rainfall map should be plotted}
}
\description{
Function which visualises daily radar rainfall depths. Requires interpolation grid and file with polygons of pixels. The radar data have been obtained from 
http://climate4impact.eu (catalog ``Radar precipitation climatology'') and are freely available.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item FileNameRadarTimeStep: Filename of radar file of rainfall depths to be visualised (NETCDF4 file).
}
}
\examples{
RainMapsRadarsDaily(20110911) # To plot daily radar rainfall map ending at 11 September 2011.

}

