% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainMapsRadarsTimeStep.R
\name{RainMapsRadarsTimeStep}
\alias{RainMapsRadarsTimeStep}
\title{Function which visualises radar rainfall depths for chosen TIMESTEP in configuration file.}
\usage{
RainMapsRadarsTimeStep()
}
\description{
Function which visualises radar rainfall depths for chosen TIMESTEP in configuration file. Requires interpolation grid and file with polygons of pixels.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
  \item FileNameRadarTimeStep: Filename of radar file of rainfall depths to be visualised (ASCII file).
}
}
\examples{
RainMapsRadarsTimeStep()

}

