% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainRetrievalMinMax.R
\name{RainRetrievalMinMax}
\alias{RainRetrievalMinMax}
\title{Function for path-averaged rainfall estimation using microwave links.}
\usage{
RainRetrievalMinMax()
}
\description{
Function for path-averaged rainfall estimation using microwave links. Works for sampling strategy where minimum and maximum received powers are provided. Maximum and minimum path-averaged rainfall intensites are computed, where a fixed correction factor is applied to remove wet antenna attenuation (subfunction ``MinMaxToMean''). Subfunction ``OutlierFilter'' can be called to apply an outlier filter.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
  \item FolderCorrected: Folder name of input data.
  \item FolderRainEstimates: Folder name of output data.
  \item OutlierFilter: is "no" when no outlier filter is to be applied; is "yes" when an outlier filter is to be applied (will only work when wet-dry classification according to nearby link approach has been performed).
}
}
\examples{
-

}

