% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainRetrievalMinMaxRSL.R
\name{RainRetrievalMinMaxRSL}
\alias{RainRetrievalMinMaxRSL}
\title{Function for path-averaged rainfall estimation using microwave links.}
\usage{
RainRetrievalMinMaxRSL(Aa = 2.3, alpha = 0.33, Data, kRPowerLawData,
  PmaxCor, PminCor, Pref)
}
\arguments{
\item{Aa}{Wet antenna attenuation correction \eqn{A_{\mbox{a}}} (dB)}

\item{alpha}{Coefficient (\eqn{\alpha}) determining contribution of minimum and 
maximum path-averaged rainfall intensity to mean path-averaged rainfall intensity (-)}

\item{Data}{Data frame with microwave link data}

\item{kRPowerLawData}{Values of coefficients a and b employed to convert specific 
attenuation to path-averaged rainfall intensity for a range of microwave frequencies}

\item{PmaxCor}{Data frame with corrected maximum received powers (dB)}

\item{PminCor}{Data frame with corrected minimum received powers (dB)}

\item{Pref}{Reference level (dB)}
}
\value{
Mean path-averaged rainfall intensity (mm h\eqn{^{-1}})
}
\description{
Function for path-averaged rainfall estimation using microwave links. 
Maximum and minimum path-averaged rainfall intensites are computed in subfunction 
``MinMaxRSLToMeanR'', where a fixed correction factor is applied to remove wet 
antenna attenuation.  

Works for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.
}
\examples{
RainRetrievalMinMaxRSL(Aa=2.3,alpha=0.33,Data=DataOutlierFiltered,
kRPowerLawData=kRPowerLawData,PmaxCor=Pcor$PmaxCor,PminCor=Pcor$PminCor,Pref=Pref)
}
\author{
Aart Overeem & Hidde Leijnse
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R. (2016): Retrieval algorithm for rainfall mapping from
microwave links in a cellular communication network, Atmospheric Measurement Techniques, under review.
}

