% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRainLocation.R
\name{ReadRainLocation}
\alias{ReadRainLocation}
\title{Function for finding (interpolated) rainfall value for a given latitude and longitude.}
\usage{
ReadRainLocation(
  dataf,
  FileGrid,
  InputCoorSystem,
  Lat,
  LocalCartesianCoorSystem,
  Lon
)
}
\arguments{
\item{dataf}{Data frame of (interpolated) rainfall values.}

\item{FileGrid}{File with interpolation grid in same coordinate system as InputCoorSystem.}

\item{InputCoorSystem}{Define EPSG code for input coordinate system (e.g., 4326L for WGS84 in degrees).}

\item{Lat}{Latitude of location for which (interpolated) rainfall value is to be extracted 
(in coordinate system InputCoorSystem).}

\item{Lot}{Longitude of location for which (interpolated) rainfall value is to be extracted 
(in coordinate system InputCoorSystem).}
}
\value{
Rainfall value for selected location (in unit of provided input rainfall data).
}
\description{
Function for finding (interpolated) rainfall value for a given latitude and longitude. 
I.e. find the grid cell which belongs to the location for which latitude and longitude are provided.
}
\examples{
ReadRainLocation(dataf=dataf,FileGrid=FileGrid,InputCoorSystem=4326L,
Lat=Lat,LocalCartesianCoorSystem=28992,Lon=Lon)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem & Hidde Leijnse
}
