% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefLevelMinMax.R
\name{RefLevelMinMax}
\alias{RefLevelMinMax}
\title{Subfunction for determination of reference signal level, which is representative of dry weather, and to obtain corrected received powers.}
\usage{
RefLevelMinMax(Pmin, PminCor, Pmin_dry, Pmax, NrStepsDay, NrStepsTwoDays,
  NrStepsDryPeriodsRefLevel, NrStepsDayMinus1, WetDry)
}
\arguments{
\item{Pmin}{minimum received power}

\item{PminCor}{corrected received power (in case of wet-dry classification corrected for dry periods)}

\item{Pmin_dry}{minimum received power}

\item{Pmax}{maximum received power}

\item{NrStepsDay}{number of time intervals in 24 hours}

\item{NrStepsTwoDays}{number of time intervals in 48 hours}

\item{NrStepsDryPeriodsRefLevel}{number of time intervals which falls in minimum number of hours that should be dry in preceding 24 hours for computing reference level}

\item{NrStepsDayMinus1}{number of time intervals in 24 ours minus 1}

\item{WetDry}{is "no" when no wet-dry classification has been applied, is "yes" when a wet-dry classification has been applied.}
}
\value{
RefLevel, PmaxCorRL, PminCorRL
}
\description{
Subfunction for determination of reference signal level, which is representative of dry weather, and to obtain corrected received powers.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
   \item HoursRefLevel: Minimum number of hours that should be dry in preceding 24 hours for computing reference level.
}
}
\examples{
-

}

