% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefLevelMinMaxRSL.R
\name{RefLevelMinMaxRSL}
\alias{RefLevelMinMaxRSL}
\title{Function for determination of reference signal level, which is representative of dry weather.}
\usage{
RefLevelMinMaxRSL(Data, Dry = NULL, HoursRefLevel = 2.5,
  PeriodHoursRefLevel = 24)
}
\arguments{
\item{Data}{Data frame with microwave link data.}

\item{Dry}{Data frame: Should interval be considered dry for reference level 
determination? (0 = wet; 1 = dry)}

\item{HoursRefLevel}{Minimum number of hours that should be dry in preceding 
PeriodHoursRefLevel hours for computing reference level (h)}

\item{PeriodHoursRefLevel}{Period over which reference level is to be determined
(h)}
}
\value{
Reference level (dB)
}
\description{
Function for determination of reference signal level, which is representative 
of dry weather.

Works for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.
}
\examples{
RefLevelMinMaxRSL(Data=DataPreprocessed,Dry=NULL,HoursRefLevel=2.5,
PeriodHoursRefLevel=24)
}
\author{
Aart Overeem & Hidde Leijnse
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R. (2016): Retrieval algorithm for rainfall mapping from
microwave links in a cellular communication network, Atmospheric Measurement Techniques, under review.
}

