% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToPolygonsRain.R
\name{ToPolygonsRain}
\alias{ToPolygonsRain}
\title{Subfunction which assignes values of rainfall grid to polygons.}
\usage{
ToPolygonsRain(Data)
}
\arguments{
\item{Data}{microwave link data}
}
\value{
Value
}
\description{
Subfunction which assignes values of rainfall grid to polygons.
}
\examples{
-

}

