% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WetDryNearbyLinkApMinMax.R
\name{WetDryNearbyLinkApMinMax}
\alias{WetDryNearbyLinkApMinMax}
\title{Function for wet-dry classification according to nearby link approach, determination of reference signal level, and corrected received powers.}
\usage{
WetDryNearbyLinkApMinMax()
}
\description{
The received signal powers often decrease during non-rainy periods, resulting in non-zero rainfall estimates, e.g., caused by reflection of the beam or dew formation on the antennas. To prevent this rainfall overestimation a reliable classification of wet and dry periods is needed. This is also beneficial for determining an appropriate reference signal level, representative for dry weather. In order to define wet and dry periods, we assume that rain is correlated in space, and hence that several links in a given area should experience a decrease in received signal level in the case of rain. A time interval is labeled as wet if at least half of the links in the vicinity (for chosen radius) of the selected link experience such a decrease. This so called nearby link approach is applied in this function. It also determines the reference signal level (subfunction RefLevelMinMax), and corrected received powers (this function and subfunction RefLevelMinMax). Further, the minimum and maximum attenuation over a link path are computed. Finally, malfunctioning link antennas can cause outliers in rainfall retrievals (especially for daily accumulations). These outliers can be removed by using a filter that is based on the assumption that rainfall is correlated in space. The value of this outlier filter is calculated, but the filter is not applied in this function.
The following parameters can be changed in the configuration file ``Config.R'':
\itemize{
  \item TIMESTEP: Duration of time interval of sampling strategy (min).
  \item FolderPreprocessed: Folder name of input data.
  \item FolderCorrected: Folder name of output data.
  \item DistanceLimit: Radius in wet-dry classification (km).
  \item MinHoursPmin: Minimum number of hours in the previous 24 hours needed for computing max(P_min) (h).
  \item ThresholdNumberLinks: Only use data if number of available links is at least larger than this threshold for present and previous day for those time steps that the selected link is available. Selected link is also counted.
  \item ThresholdMedianL: Threshold value (dB km^-1).
  \item ThresholdMedian: Threshold value (dB).
  \item FileRainRetr: Filename with values of coefficients in relationship between specific attenuation and rainfall intensity.
}
}
\examples{
-

}

