% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntpPathToPoint.R
\name{IntpPathToPoint}
\alias{IntpPathToPoint}
\title{Subfunction for computing path-averaged rainfall intensities for unique link paths. 
A path-averaged rainfall intensity is assigned to a point at the middle of 
the link path.}
\usage{
IntpPathToPoint(ID, Rmean, XEnd, XStart, YEnd, YStart)
}
\arguments{
\item{ID}{Link identifier.}

\item{Rmean}{Data frame with mean path-averaged rainfall intensities (mm h\eqn{^{-1}}).}

\item{XEnd}{Easting of end of links (km).}

\item{XStart}{Easting of start of links (km).}

\item{YEnd}{Northing of end of links (km).}

\item{YStart}{Northing of start of links (km).}
}
\value{
Coordinates of links in Azimuthal Equidistant Cartesian coordinate system
(latitude, longitude) and rainfall intensity (mm h\eqn{^{-1}})).
}
\description{
Subfunction for computing path-averaged rainfall intensities for 
unique link paths. The link-based, e.g. a 15-minute path-averaged rainfall 
accumulation is converted to a path-averaged rainfall intensity, and subsequently 
assigned to a point at the middle of the link path. Path-averaged rainfall 
intensities are obtained, so data from full-duplex links are averaged.
}
\examples{
IntpPathToPoint(ID=ID,Rmean=Rmean,Xend=Xend,XStart=XStart,YEnd=YEnd,YStart=YStart)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem & Hidde Leijnse
}
