% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathLength.R
\name{PathLength}
\alias{PathLength}
\title{Function to compute path length of microwave links from coordinates.}
\usage{
PathLength(XStart, XEnd, YStart, YEnd)
}
\arguments{
\item{XStart}{List with longitude of start of microwave links (decimal degrees)}

\item{XEnd}{List with longitude of end of microwave links (decimal degrees)}

\item{YStart}{List with latitude of start of microwave links (decimal degrees)}

\item{YEnd}{List with latitude of end of microwave links (decimal degrees)}
}
\value{
Path length of microwave links (km).
}
\description{
Function for computing path length of microwave links from coordinates.

Coordinates should be provided as longitudes and latitudes of start and end of a link.
It uses the function ''distVincentyEllipsoid'' from the R package ''geosphere'', which computes the ''Vincenty'' (Ellipsoid) Great Circle Distance.
The shortest distance between two points (i.e., the ''great-circle-distance'' or ''as the crow flies''), according to the ''Vincenty (ellipsoid)'' method. 
The WGS84 ellipsoid is used.
}
\examples{
PathLength(XStart=XStart,XEnd=XEnd,YStart=YStart,YEnd=YEnd)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem
}
