% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLinkLocations.R
\name{PlotLinkLocations}
\alias{PlotLinkLocations}
\title{Function which visualises microwave link paths on a map.}
\usage{
PlotLinkLocations(
  AlphaLinkLocations,
  BBoxOSMauto,
  OSMBottom,
  ColourLinks,
  ColourType,
  dataf,
  DateTime,
  ExtraTextLinkLocations,
  FigFileLinkLocations,
  FigHeight,
  FigWidth,
  FilePolygonsGrid,
  FolderFigures,
  FontFamily,
  GoogleLocDegSpecified,
  GoogleLocLat,
  GoogleLocLon,
  GoogleLocName,
  GoogleLocNameSpecified,
  GoogleMapType,
  GoogleZoomlevel,
  LabelAxisLat,
  LabelAxisLonGoogle,
  LabelAxisLonOSM,
  LabelAxisLonStamen,
  MapBackground,
  OSMLeft,
  OSMRight,
  OSMScale,
  OSMTop,
  OutputFileType,
  SizeLinks,
  SizePlotTitle,
  StamenMapType,
  StamenZoomlevel,
  TitleLinkLocations
)
}
\arguments{
\item{AlphaLinkLocations}{Transparency of link paths.}

\item{BBoxOSMauto}{Compute bounding box from input data or used bounding box defined above?
(for OpenStreetMap and Stamen Map only). Use "yes" if bounding box is to be computed from interpolation grid.}

\item{OSMBottom}{Latitude in degrees (WGS84) for bottom side of the area for which rainfall depths 
are to be plotted (for OpenStreetMap & Stamen Maps only).}

\item{ColourLinks}{Colour of plotted link paths.}

\item{ColourType}{Colour or black-and-white background map? Use "color" for colour and "bw" 
for black-and-white background map.}

\item{dataf}{data frame which contains (at least) locations of microwave links in Azimuthal 
Equidistant Cartesian coordinate system.}

\item{DateTime}{Date and time for which link locations are plotted. This is used in the
title caption of the figure and in the file name.}

\item{ExtraTextLinkLocations}{Second part of title of plot.}

\item{FigFileLinkLocations}{Part of figure output file name.}

\item{FigHeight}{Figure height. 1280 times 1280 pixels seems maximum graphical resolution for 
downloaded Google Maps. Because also axes and legend are plotted, it is advised to use e.g. 
1450 times 1450 pixels. 
Then the Google Map will remain 
approximately 1280 times 1280 pixels. Using higher values is not a problem (e.g. 2000). 
In this way it is tried to get the highest possible resolution. For OpenStreetMap the maps may 
reach resolutions of 1500 - 2000 pixels. Hence, using FigWidth and FigHeight of 2000 pixels or 
higher is advised. The OpenStreetMap itself is stored in file "ggmapTemp.png". From this file 
the resolution of the background map can be obtained. This can be useful for determining an 
appropriate FigWidth and FigHeight above.}

\item{FigWidth}{Figure width. 1280 times 1280 pixels seems maximum graphical resolution for 
downloaded Google Maps. Because also axes and legend are plotted, it is advised to use e.g. 
1450 times 1450 pixels. Then the Google Map will remain approximately 1280 times 1280 pixels. 
Using higher values is not a problem (e.g. 2000). In this way it is tried to get the 
highest possible resolution. For OpenStreetMap the maps may reach resolutions of 1500 - 2000 
pixels. Hence, using FigWidth and FigHeight of 2000 pixels or higher is advised. The OpenStreetMap 
itself is stored in file "ggmapTemp.png". From this file the resolution of the background map can 
be obtained. This can be useful for determining an appropriate FigWidth and FigHeight above.}

\item{FilePolygonsGrid}{Name of file with polygons of interpolation grid.}

\item{FolderFigures}{Folder name of figures.}

\item{FontFamily}{Specify font family of text in figures. To select the default font use "". 
Using "Times" may give warnings when executing the visualisation. In that case the font is not 
installed on the computer. This can be solved by using the default font ("").}

\item{GoogleLocDegSpecified}{If GoogleLocDegSpecified is "yes" then the specified location in 
degrees is used is used as the centre of the Google Map. If both GoogleLocNameSpecified and 
GoogleLocDegSpecified are not equal to "yes", the bounding box of the map is determined from 
the provided grid and used as centre of the Google Map.}

\item{GoogleLocLat}{Latitude of middle of Google Map (degrees).}

\item{GoogleLocLon}{Longitude of middle of Google Map (degrees).}

\item{GoogleLocName}{Location of middle of Google Map, provided as text, e.g. name of city,
street name, country.}

\item{GoogleLocNameSpecified}{If GoogleLocNameSpecified is "yes" then the specified location 
name GoogleLocName is used as the centre of the Google Map. If both GoogleLocNameSpecified and 
GoogleLocDegSpecified are not equal to "yes", the bounding box of the map is determined from the 
provided grid and used as centre of the Google Map.}

\item{GoogleMapType}{In case of Google Maps: which map type should be used? Available map types: 
"terrain", "satellite", "roadmap", and "hybrid".}

\item{GoogleZoomlevel}{Which zoom level to use for the Google Maps?}

\item{LabelAxisLat}{Label name of vertical axis.}

\item{LabelAxisLonGoogle}{Label name of horizontal axis (for Google Maps only).}

\item{LabelAxisLonOSM}{Label name of horizontal axis (for OpenStreetMap only).}

\item{LabelAxisLonStamen}{Label name of horizontal axis (for Stamen Map only).}

\item{MapBackground}{Google Maps, OpenStreetMap or Stamen Map as background? Use "Google" for Google Maps, 
"OSM" for OpenStreetMap and "Stamen" for Stamen Map (based on OpenStreetMap data). 
Note that Google Maps will only plot on a square figure.
It seems that mapping with OpenStreetMap (“get openstreetmap”) is no langer supported.
This implies that mapping can only be done employing Google Maps (if Google API key is obtained) or via
Stamen Map. This is not related to the RAINLINK version.}

\item{OSMLeft}{Longitude in degrees (WGS84) for left side of the area for which rainfall depths are 
to be plotted (for OpenStreetMap & Stamen Maps only).}

\item{OSMRight}{Longitude in degrees (WGS84) for right side of the area for which rainfall depths 
are to be plotted (for OpenStreetMap & Stamen Maps only).}

\item{OSMScale}{Give value of scale (for OpenStreetMap only). A proper choice of the scale parameter 
in get_openstreetmap is difficult. It cannot be computed automatically. Hence, a scale parameter value 
should be provided below. The scale parameter should be as small as possible to get the highest 
graphical resolution. However, a too low value may result in a map not being downloaded. Hence, the user 
should manually supply get_openstreetmap with a scale. It may require some iterations to find the 
appropriate value for scale. The file "ggmapTemp.png" is written to disk when an OpenStreetMap is loaded. 
The highest possible resolution for a square area is about 2000 x 2000 pixels.}

\item{OSMTop}{Latitude in degrees (WGS84) for top side of the area for which rainfall depths are to be 
plotted (for OpenStreetMap & Stamen Maps only).}

\item{OutputFileType}{Choose output file type of image: jpeg, png or tiff.}

\item{SizeLinks}{Size of plotted link paths.}

\item{SizePlotTitle}{Size of plot title.}

\item{StamenMapType}{In case of Stamen Maps: which map type should be used? Available map types which 
seem most useful and work: "toner-hybrid" &, recommended: "toner-lite", "terrain" & "watercolor".}

\item{StamenZoomlevel}{Which zoom level to use for the Stamen Maps? This determines the level of detail. 
Large values take more time. It does not determine the domain of the area which is plotted.}

\item{TitleLinkLocations}{First part of title of plot.}
}
\description{
Function which visualises microwave link paths on a map.
}
\examples{
PlotLinkLocations(AlphaLinkLocations=AlphaLinkLocations,BBoxOSMauto=BBoxOSMauto,
OSMBottom=OSMBottom,ColourLinks=ColourLinks,ColourType=ColourType,dataf=dataf,
DateTime=DateTime,ExtraTextLinkLocations=ExtraTextLinkLocations,
FigFileLinkLocations=FigFileLinkLocations,FigHeight=FigHeight,
FigWidth=FigWidth,FilePolygonsGrid=FilePolygonsGrid,FolderFigures=FolderFigures,
FontFamily=FontFamily,GoogleLocDegSpecified=GoogleLocDegSpecified,
GoogleLocLat=GoogleLocLat,GoogleLocLon=GoogleLocLon,GoogleLocName=GoogleLocName,
GoogleLocNameSpecified=GoogleLocNameSpecified,GoogleMapType=GoogleMapType,
GoogleZoomlevel=GoogleZoomlevel,LabelAxisLat=LabelAxisLat,
LabelAxisLonGoogle=LabelAxisLonGoogle,LabelAxisLonOSM=LabelAxisLonOSM,
LabelAxisLonStamen=LabelAxisLonStamen,MapBackground=MapBackground,OSMLeft=OSMLeft,
OSMRight=OSMRight,OSMScale=OSMScale,OSMTop=OSMTop,OutputFileType=OutputFileType,
SizeLinks=SizeLinks,SizePlotTitle=SizePlotTitle,StamenMapType=StamenMapType,
StamenZoomlevel=StamenZoomlevel,TitleLinkLocations=TitleLinkLocations)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem & Hidde Leijnse
}
