% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygons.R
\name{Polygons}
\alias{Polygons}
\title{Subfunction which makes dataframe for polygons with rainfall estimates in specific rainfall class.}
\usage{
Polygons(cond, Selected)
}
\arguments{
\item{cond}{Row numbers of dataframe which fall in specific rainfall class}

\item{Selected}{Coordinates of polygons and their assigned rainfall values}
}
\value{
Data frame
}
\description{
Subfunction which makes dataframe for polygons with rainfall estimates in specific rainfall class.
}
\examples{
Polygons(cond=cond,Selected=Selected)
}
\author{
Aart Overeem & Hidde Leijnse
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R. (2016): Retrieval algorithm for rainfall mapping from
microwave links in a cellular communication network, Atmospheric Measurement Techniques, under review.
}

