% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingMinMaxRSL.R
\name{PreprocessingMinMaxRSL}
\alias{PreprocessingMinMaxRSL}
\title{Function for preprocessing of microwave link data.}
\usage{
PreprocessingMinMaxRSL(
  Data,
  MaxFrequency = Inf,
  MinFrequency = 0,
  verbose = TRUE
)
}
\arguments{
\item{Data}{Data frame with microwave link data (use data(Linkdata) to load example data).}

\item{MaxFrequency}{Maximum allowed microwave frequency of link in output (GHz; default infinite).}

\item{MinFrequency}{Minimum allowed microwave frequency of link in output (GHz; default 0).}
}
\value{
Data frame with microwave link data.
}
\description{
Function for preprocessing of microwave link data. This function performs the following tasks:  
\enumerate{
  \item Link data are selected for microwave frequencies within chosen range.
  \item Data selection criteria are applied.
}

Works for a sampling strategy where minimum and maximum received signal powers
are provided, and the transmitted power levels are constant.

Also works for a sampling strategy where instantaneous transmitted and received signal levels are obtained.
In case of instantaneous signal levels, it does not matter whether transmitted power levels vary or are constant.
The only requirement is that the input data for RAINLINK needs some preprocessing. See ''ManualRAINLINK.pdf''
for instructions. 

The input microwave link data do not have to be sorted chronologically.

It is strongly advised to use the same unique link identifier (ID) for a link during
the entire processed period(s). First of all, time series of sufficient length
are needed in order to compute e.g. a reference signal level. Moreover, utilizing
the same ID allows for plotting (continuous) time series from the same link.
}
\examples{
data(Linkdata)
PreprocessingMinMaxRSL(Data=Linkdata,MaxFrequency=40.5,MinFrequency=12.5)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem & Hidde Leijnse & Lotte de Vos
}
