% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainMapsLinksDaily.R
\name{RainMapsLinksDaily}
\alias{RainMapsLinksDaily}
\title{Function which visualises daily link-based rainfall depths.}
\usage{
RainMapsLinksDaily(
  AlphaLinksDaily,
  AlphaPlotLocation,
  AlphaPolygon,
  AutDefineLegendTop,
  BBoxOSMauto,
  ColourLinks,
  ColoursNumber,
  ColourPlotLocation,
  ColourPlotLocationText,
  ColourScheme,
  ColourType,
  ColourHighestClass,
  ConversionDepthToIntensity,
  CoorSystemInputData,
  DateTimeEndRainMaps,
  DateTimeStartRainMaps,
  ExtraDeg,
  ExtraText,
  FigFileLinksDaily,
  FigHeight,
  FigWidth,
  FileGrid,
  FilePolygonsGrid,
  FolderFigures,
  FolderRainMaps,
  FolderRainEstimates,
  FontFamily,
  GoogleLocDegSpecified,
  GoogleLocLat,
  GoogleLocLon,
  GoogleLocName,
  GoogleLocNameSpecified,
  GoogleMapType,
  GoogleZoomlevel,
  LabelAxisLat,
  LabelAxisLonGoogle,
  LabelAxisLonOSM,
  LabelAxisLonStamen,
  LatLocation,
  LatText,
  LegendSize,
  LegendTitleLinksDaily,
  LonLocation,
  LonText,
  ManualScale,
  MapBackground,
  OSMBottom,
  OSMLeft,
  OSMRight,
  OSMScale,
  OSMTop,
  OutputFileType,
  PERIOD,
  PlotLocation,
  PixelBorderCol,
  PlotBelowScaleBottom,
  PlotLocLinks,
  ScaleBottomDaily,
  ScaleHigh,
  ScaleLow,
  ScaleTopDaily,
  SizeLinks,
  SizePixelBorder,
  SizePlotLocation,
  SizePlotTitle,
  StamenMapType,
  StamenZoomlevel,
  SymbolPlotLocation,
  TIMESTEP,
  TitleLinks,
  XMiddle,
  YMiddle
)
}
\arguments{
\item{AlphaLinksDaily}{Transparency of link paths.}

\item{AlphaPlotLocation}{Transparency of plotted symbol for specified location on map.}

\item{AlphaPolygon}{Transparency of polygons.}

\item{AutDefineLegendTop}{Let R automatically define highest value of legend in case of "yes". 
Then the highest class, i.e. the one plotted separately above the other classes, is not plotted anymore.}

\item{BBoxOSMauto}{Compute bounding box from input data or used bounding box defined above? 
(for OpenStreetMap and Stamen Map only). Use "yes" if bounding box is to be computed from interpolation grid.}

\item{ColourLinks}{Colour of plotted link paths.}

\item{ColoursNumber}{Number of colour classes in legend.}

\item{ColourPlotLocation}{Colour of plotted symbol for specified location on map.}

\item{ColourPlotLocationText}{Colour of plotted rainfall depth for specified location on map.}

\item{ColourScheme}{Hexadecimal codes or names for colours of legend.}

\item{ColourType}{Colour or black-and-white background map? Use "color" for colour and "bw" for 
black-and-white background map.}

\item{ColourHighestClass}{Colour of highest class.}

\item{ConversionDepthToIntensity}{Conversion factor from rainfall depth (mm) to intensity (mm/h).}

\item{CoorSystemInputData}{Define coordinate system of input data (e.g. "+init=epsg:4326" for 
WGS84 in degrees).}

\item{DateTimeEndRainMaps}{Date and time at which rainfall mapping ends.}

\item{DateTimeStartRainMaps}{Date and time at which rainfall mapping starts.}

\item{ExtraDeg}{To reduce computational time, it is automatically determined which grid cells fall 
within the plotted region. 
To also plot grid cell values which partly fall outside the plotted region, a positive number for 
ExtraDeg should be specified (degrees). 
This should typically be at least the size of one grid cell in degrees.}

\item{ExtraText}{Second part of title of plot.}

\item{FigFileLinksDaily}{Part of figure output file name.}

\item{FigHeight}{Figure height. 1280 times 1280 pixels seems maximum graphical resolution for 
downloaded Google Maps. Because also axes and legend are plotted, it is advised to use e.g. 1450 
times 1450 pixels. Then the Google Map will remain approximately 1280 times 1280 pixels. Using higher 
values is not a problem (e.g. 2000). In this way it is tried to get the highest possible resolution. 
For OpenStreetMap the maps may reach resolutions of 1500 - 2000 pixels. Hence, using FigWidth 
and FigHeight of 2000 pixels or higher is advised. The OpenStreetMap itself is stored in file 
"ggmapTemp.png". From this file the resolution of the background map can be obtained. This can be useful 
for determining an appropriate FigWidth and FigHeight above.}

\item{FigWidth}{Figure width. 1280 times 1280 pixels seems maximum graphical resolution for downloaded 
Google Maps. Because also axes and legend are plotted, it is advised to use e.g. 1450 times 1450 pixels. 
Then the Google Map will remain approximately 1280 times 1280 pixels. Using higher values is not a problem 
(e.g. 2000). In this way it is tried to get the highest possible resolution. For OpenStreetMap the maps may 
reach resolutions of 1500 - 2000 pixels. Hence, using FigWidth and FigHeight of 2000 pixels or higher is 
advised. The OpenStreetMap itself is stored in file "ggmapTemp.png". From this file the resolution of the 
background map can be obtained. This can be useful for determining an appropriate FigWidth and FigHeight 
above.}

\item{FileGrid}{File with interpolation grid in same coordinate system as CoorSystemInputData.}

\item{FilePolygonsGrid}{Name of file with polygons of interpolation grid.}

\item{FolderFigures}{Folder name of figures.}

\item{FolderRainMaps}{Folder name of interpolated link data (input).}

\item{FolderRainEstimates}{Folder name of input link path data.}

\item{FontFamily}{Specify font family of text in figures. To select the default font use "". 
Using "Times" may give warnings when executing the visualisation. In that case the font is not installed 
on the computer. This can be solved by using the default font ("").}

\item{GoogleLocDegSpecified}{If GoogleLocDegSpecified is "yes" then the specified location in degrees is 
used as the centre of the Google Map. If both GoogleLocNameSpecified and GoogleLocDegSpecified are not 
equal to "yes", the bounding box of the map is determined from the provided grid and used as centre of the 
Google Map.}

\item{GoogleLocLat}{Latitude of middle of Google Map (degrees).}

\item{GoogleLocLon}{Longitude of middle of Google Map (degrees).}

\item{GoogleLocName}{Location of middle of Google Map, provided as text, e.g. name of city, street name, 
country.}

\item{GoogleLocNameSpecified}{If GoogleLocNameSpecified is "yes" then the specified location name 
GoogleLocName is used as the centre of the Google Map. If both GoogleLocNameSpecified and 
GoogleLocDegSpecified are not equal to "yes", the bounding box of the map is determined from the 
provided grid and used as centre of the Google Map.}

\item{GoogleMapType}{In case of Google Maps: which map type should be used? Available map types: 
"terrain", "satellite", "roadmap", and "hybrid".}

\item{GoogleZoomlevel}{Which zoom level to use for the Google Maps?}

\item{LabelAxisLat}{Label name of vertical axis.}

\item{LabelAxisLonGoogle}{Label name of horizontal axis (for Google Maps only).}

\item{LabelAxisLonOSM}{Label name of horizontal axis (for OpenStreetMap only).}

\item{LabelAxisLonStamen}{Label name of horizontal axis (for Stamen Map only).}

\item{LatLocation}{Latitude of location on map (degrees).}

\item{LatText}{Latitude of text (rainfall depth) of location on map (degrees).}

\item{LegendSize}{Size of legend (choose e.g. 75 for 6 classes and 50 for 10 classes).}

\item{LegendTitleLinksDaily}{Title of legend.}

\item{LonLocation}{Longitude of location on map (degrees).}

\item{LonText}{Longitude of text (rainfall depth) of location on map (degrees).}

\item{ManualScale}{Manually supply the legend breaks if ManuelScale is not equal to "no". 
Interval breaks are determined manually from ScaleLow and ScaleHigh. If ManualScale is "no" 
interval breaks are determined automatically.}

\item{MapBackground}{Google Maps, OpenStreetMap or Stamen Map as background? Use "Google" for Google Maps, 
"OSM" for OpenStreetMap and "Stamen" for Stamen Map (based on OpenStreetMap data). 
Note that Google Maps will only plot on a square figure.
It seems that mapping with OpenStreetMap (“get openstreetmap”) is no langer supported.
This implies that mapping can only be done employing Google Maps (if Google API key is obtained) or via
Stamen Map. This is not related to the RAINLINK version.}

\item{OSMBottom}{Latitude in degrees (WGS84) for bottom side of the area for which rainfall depths 
are to be plotted (for OpenStreetMap & Stamen Maps only).}

\item{OSMLeft}{Longitude in degrees (WGS84) for left side of the area for which rainfall depths 
are to be plotted (for OpenStreetMap & Stamen Maps only).}

\item{OSMRight}{Longitude in degrees (WGS84) for right side of the area for which rainfall depths 
are to be plotted (for OpenStreetMap & Stamen Maps only).}

\item{OSMScale}{Give value of scale (for OpenStreetMap only). A proper choice of the scale parameter 
in get_openstreetmap is difficult. It cannot be computed automatically. Hence, a scale parameter value 
should be provided below. The scale parameter should be as small as possible to get the highest 
graphical resolution. However, a too low value may result in a map not being downloaded. Hence, 
the user should manually supply get_openstreetmap with a scale. It may require some iterations to find 
the appropriate value for scale. The file "ggmapTemp.png" is written to disk when an OpenStreetMap is 
loaded. The highest possible resolution for a square area is about 2000 x 2000 pixels.}

\item{OSMTop}{Latitude in degrees (WGS84) for top side of the area for which rainfall depths are to be 
plotted (for OpenStreetMap & Stamen Maps only).}

\item{OutputFileType}{Choose output file type of image: jpeg, png or tiff.}

\item{PERIOD}{Select daily time interval, i.e. "0800" implies 0800 UTC previous day - 0800 UTC present 
day (use 2400 for 0000 UTC).}

\item{PlotLocation}{A location is plotted on map if PlotLocation is "yes".}

\item{PixelBorderCol}{Choose colour of pixel borders. Use NA (without quotes) to not plot pixel borders. 
If the pixels are relatively small with respect to the plotted region, 
the graphical quality of the pixel borders deteriorates due to low number of pixels (low resolution).}

\item{PlotBelowScaleBottom}{Plot grid lines for polygons below threshold ScaleBottomTimeStep or 
ScaleBottomDaily? If "yes" grid lines are plotted, otherwise they are not plotted.}

\item{PlotLocLinks}{Plot locations of links in plot? If "yes" than locations of links are plotted in the 
plot. Note that full-duplex links are plotted twice.}

\item{ScaleBottomDaily}{Lowest class starts at this threshold (minimum rainfall accumulation (mm) to be 
plotted). Using a value clearly above 0 mm can save a lot of computation time if the polygons belonging 
to values below the threshold are not plotted.}

\item{ScaleHigh}{ScaleHigh Highest value per class interval, i.e. the highest legend breaks, if these 
are manually chosen.  Please note that in case of x values in ColoursNumber, ScaleHigh should also 
contain x values.}

\item{ScaleLow}{ScaleLow Lowest value per class interval, i.e. the lowest legend breaks, if these are 
manually chosen. Please note that in case of x values in ColoursNumber, ScaleLow should also contain x 
values.}

\item{ScaleTopDaily}{Highest colour class ends here (maximum rainfall accumulation (mm) to be plotted). 
Sometimes the legend is not correctly plotted. In that case try other values for ScaleTopDaily and/or 
ScaleBottomDaily For instance, if the highest class (> x mm) is plotted below instead of above the other 
classes. Or if the number of classes does not match the number of chosen classes. Another way to prevent 
this is to manually give the legend breaks (ManualScale not equal to "no").}

\item{SizeLinks}{Size of plotted link paths.}

\item{SizePixelBorder}{Size of pixel borders.}

\item{SizePlotLocation}{Size of symbol and and accompanied text for specified location on map.}

\item{SizePlotTitle}{Size of plot title.}

\item{StamenMapType}{In case of Stamen Maps: which map type should be used? Available map types which 
seem most useful and work: "toner-hybrid" &, recommended: "toner-lite", "terrain" & "watercolor".}

\item{StamenZoomlevel}{Which zoom level to use for the Stamen Maps? This determines the level of detail. 
Large values take more time. It does not determine the domain of the area which is plotted.}

\item{SymbolPlotLocation}{Symbol to be plotted for specified location on map.}

\item{TIMESTEP}{Duration of time interval of sampling strategy (min).}

\item{TitleLinks}{First part of title of plot.}

\item{XMiddle}{The longitude of the centre of the Azimuthal Equidistant Cartesian coordinate system, 
given in the coordinate system of the input data.}

\item{YMiddle}{The latitude of the centre of the Azimuthal Equidistant Cartesian coordinate system, 
given in the coordinate system of the input data.}
}
\description{
Function which visualises daily link-based rainfall depths. Requires interpolation grid and 
file with polygons of pixels. Daily rainfall depths are computed irrespective of the number of available 
files. If, for instance, only one out of 96 files is available, the daily rainfall depth is still computed 
and visualised. The data availability is plotted as a percentage in the title caption of the graph.
Function will also plot accumulations for other aggregation intervals. Note that the data availability,
plotted in the figure caption is only correctly computed for daily intervals.
}
\examples{
RainMapsLinksDaily(AlphaLinksDaily=AlphaLinksDaily,AlphaPlotLocation=AlphaPlotLocation,
AlphaPolygon=AlphaPolygon,AutDefineLegendTop=AutDefineLegendTop,BBoxOSMauto=BBoxOSMauto,
ColourLinks=ColourLinks,ColoursNumber=ColoursNumber,
ColourPlotLocation=ColourPlotLocation,ColourPlotLocationText=ColourPlotLocationText,
ColourScheme=ColourScheme,ColourType=ColourType,
ConversionDepthToIntensity=ConversionDepthToIntensity,
CoorSystemInputData=CoorSystemInputData,DateTimeEndRainMaps=DateTimeEndRainMaps,
DateTimeStartRainMaps=DateTimeStartRainMaps,ExtraDeg=ExtraDeg,ExtraText=ExtraText,
FigFileLinksDaily=FigFileLinksDaily,FigHeight=FigHeight,FigWidth=FigWidth,
FileGrid=FileGrid,FilePolygonsGrid=FilePolygonsGrid,FolderFigures=FolderFigures,
FolderRainMaps=FolderRainMaps,FolderRainEstimates=FolderRainEstimates,
FontFamily=FontFamily,GoogleLocDegSpecified=GoogleLocDegSpecified,
GoogleLocLat=GoogleLocLat,GoogleLocLon=GoogleLocLon,GoogleLocName=GoogleLocName,
GoogleLocNameSpecified=GoogleLocNameSpecified,GoogleMapType=GoogleMapType,
GoogleZoomlevel=GoogleZoomlevel,LabelAxisLat=LabelAxisLat,
LabelAxisLonGoogle=LabelAxisLonGoogle,LabelAxisLonOSM=LabelAxisLonOSM,
LabelAxisLonStamen=LabelAxisLonStamen,LatLocation=LatLocation,LatText=LatText,
LegendSize=LegendSize,LegendTitleLinksDaily=LegendTitleLinksDaily,
LonLocation=LonLocation,LonText=LonText,ManualScale=ManualScale,
MapBackground=MapBackground,OSMBottom=OSMBottom,OSMLeft=OSMLeft,
OSMRight=OSMRight,OSMScale=OSMScale,OSMTop=OSMTop,OutputFileType=OutputFileType,
PERIOD=PERIOD,PlotLocation=PlotLocation,PixelBorderCol=PixelBorderCol,
PlotBelowScaleBottom=PlotBelowScaleBottom,PlotLocLinks=PlotLocLinks,
ScaleBottomDaily=ScaleBottomDaily,ScaleHigh=ScaleHigh,ScaleLow=ScaleLow,
ScaleTopDaily=ScaleTopDaily,SizeLinks=SizeLinks,SizePixelBorder=SizePixelBorder,
SizePlotLocation=SizePlotLocation,SizePlotTitle=SizePlotTitle,
StamenMapType=StamenMapType,StamenZoomlevel=StamenZoomlevel,
SymbolPlotLocation=SymbolPlotLocation,TIMESTEP=TIMESTEP,TitleLinks=TitleLinks,
XMiddle=XMiddle,YMiddle=YMiddle)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem & Hidde Leijnse
}
