% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topology.R
\name{Topology}
\alias{Topology}
\title{Function for plotting characteristics of topology of commercial microwave link network.}
\usage{
Topology(
  Data,
  CoorSystemInputData = NULL,
  FigNameBarplotAngle,
  FigNameBarplotFrequency,
  FigNameBarplotPathLength,
  FigNameFrequencyVsPathLength,
  FigNameScatterdensityplotFrequencyVsPathLength,
  Maxf,
  Minf,
  MaxL,
  MinL,
  Rmean = NULL,
  Stepf,
  StepL,
  verbose = TRUE
)
}
\arguments{
\item{Data}{Data frame with microwave link data (use data(Linkdata) to load example data).}

\item{CoorSystemInputData}{Define coordinate system of input data (in case of
WGS84 provide NULL).}

\item{FigNameBarplotAngle}{Name of file with bar plot with percentage of links for bins of link orientation. The extension must be ''.pdf''.}

\item{FigNameBarplotFrequency}{Name of file with bar plot with percentage of links for bins of link path length. The extension must be ''.pdf''.}

\item{FigNameBarplotPathLength}{Name of file with bar plot with percentage of links for bins of microwave frequency. The extension must be ''.pdf''.}

\item{FigNameFrequencyVsPathLength}{Name of file with scatter plot of microwave frequency versus link path length. The extension must be ''.pdf''.}

\item{FigNameScatterdensityplotFrequencyVsPathLength}{Name of file with scatter density plot of microwave frequency versus link path length. The extension must be ''.pdf''.}

\item{Maxf}{Maximum microwave frequency to be plotted in bar plot (GHz). This is the value where the last bin class ends.}

\item{Minf}{Minimum microwave frequency to be plotted in bar plot (GHz). This is the value where the first bin class ends.}

\item{MaxL}{Maximum link path length to be plotted in bar plot (km). This is the value where the last bin class ends.}

\item{MinL}{Minimum link path length to be plotted in bar plot (km). This is the value where the first bin class ends.}

\item{Rmean}{Vector of link-derived rainfall intensities (mm h\eqn{^{-1}}) with length equal to Data.}

\item{Stepf}{Bin size of microwave frequency classes for bar plot in GHz.}

\item{StepL}{Bin size of link path length classes for bar plot in km.}
}
\value{
Figures with characteristics of topology of commercial microwave link network.
}
\description{
Function for plotting characteristics of topology of commercial microwave link data. This function provides the following figures:  
\enumerate{
  \item Bar plot with percentage of links for bins of link path length (km).
  \item Bar plot with percentage of links for bins of microwave frequency (GHz).
  \item Bar plot with percentage of links for bins of orientation (degrees).
  \item Scatter plot of microwave frequency (GHz) versus link path length (km).
  \item Scatter density plot of microwave frequency (GHz) versus link path length (km).
}

Does not depend on sampling strategy.

The input microwave link data do not have to be sorted chronologically. Full-duplex links will give two data entries, these will both be used. 

The computed percentages in the bar plots are only based on the range of classes presented in the bar plots, i.e. data outside these classes are not used in the computations.

When Rmean is provided, all figures are only based on data where the link-derived rainfall intensities are equal to or larger than 0 mm h\eqn{^{-1}}.
}
\examples{
data(Linkdata)
Topology(Data=Linkdata,CoorSystemInputData=NULL,FigNameBarplotAngle="Barplot_Orientation.pdf",
FigNameBarplotFrequency="Barplot_Frequency.pdf",FigNameBarplotPathLength="Barplot_PathLength.pdf",
FigNameFrequencyVsPathLength="Frequency_vs_PathLength.pdf",
FigNameScatterdensityplotFrequencyVsPathLength="ScatterdensityPlot_Frequency_vs_PathLength.pdf",
Maxf=40,Minf=13,MaxL=21,MinL=1,Rmean=Rmean,Stepf=1.5,StepL=2)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem
}
