% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToPolygonsRain.R
\name{ToPolygonsRain}
\alias{ToPolygonsRain}
\title{Subfunction which assignes values of rainfall grid to polygons.}
\usage{
ToPolygonsRain(Data)
}
\arguments{
\item{Data}{Field of rainfall depths at the chosen grid.}
}
\value{
Field of rainfall depths for the polygons at the chosen grid.
}
\description{
Subfunction which assignes values of rainfall grid to polygons.
}
\examples{
ToPolygonsRain(Data=Data)
}
\references{
''ManualRAINLINK.pdf''

Overeem, A., Leijnse, H., and Uijlenhoet, R., 2016: Retrieval algorithm for rainfall mapping from microwave links in a 
cellular communication network, Atmospheric Measurement Techniques, 9, 2425-2444, https://doi.org/10.5194/amt-9-2425-2016.
}
\author{
Aart Overeem & Hidde Leijnse
}
